finish section-5

+++ src/items/items.service.ts
@@ -1,9 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ItemStatus } from './item-status.enum';
 import { CreateItemDto } from './dto/create-item.dto';
 import { Item } from 'src/entities/item.entity';
 import { Repository } from 'typeorm';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ItemsService {
@@ -23,7 +28,7 @@ export class ItemsService {
     return item;
   }
 
-  async create(creatteItemDto: CreateItemDto) {
+  async create(creatteItemDto: CreateItemDto, user: User) {
     const { name, price, description } = creatteItemDto;
     const item = this.itemRepository.create({
       name,
@@ -32,13 +37,17 @@ export class ItemsService {
       status: ItemStatus.ON_SALE,
       createdAt: new Date().toISOString(),
       updatedAt: new Date().toISOString(),
+      user,
     });
     await this.itemRepository.save(item);
     return item;
   }
 
-  async updateStatus(id: string) {
+  async updateStatus(id: string, user: User) {
     const item = await this.findById(id);
+    if (item.userId === user.id) {
+      throw new BadRequestException('自身の商品を購入することはできません');
+    }
     item.status = ItemStatus.SOLD_OUT;
     item.updatedAt = new Date().toISOString();
     const updatedItem = await this.itemRepository.update(id, {
@@ -51,8 +60,13 @@ export class ItemsService {
     return item;
   }
 
-  async deleteById(id: string) {
+  async deleteById(id: string, user: User) {
+    const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません');
+    }
     const response = await this.itemRepository.delete({ id });
+
     if (response.affected !== 1) {
       throw new NotFoundException(`${id}のデータを削除できませんでした`);
     }
